+++ apps/backend/src/lib/webhooks.tsx
@@ -11,7 +11,7 @@ async function sendWebhooks(options: {
   data: any,
 }) {
   const apiKey = getEnvVariable("STACK_SVIX_API_KEY");
-  const server = getEnvVariable("STACK_SVIX_SERVER_URL", undefined);
+  const server = getEnvVariable("STACK_SVIX_SERVER_URL", "") || undefined;
   const svix = new Svix(apiKey, { serverUrl: server });
 
   await svix.application.getOrCreate({ uid: options.projectId, name: options.projectId });