+++ apps/dashboard/sentry.edge.config.ts
@@ -4,6 +4,7 @@
 // https://docs.sentry.io/platforms/javascript/guides/nextjs/
 
 import * as Sentry from "@sentry/nextjs";
+import { nicify } from "@stackframe/stack-shared/dist/utils/strings";
 
 Sentry.init({
   dsn: "https://6e618f142965a385267f1030793e0400@o4507084192022528.ingest.us.sentry.io/4507084192219136",
@@ -15,4 +16,26 @@ Sentry.init({
   debug: false,
 
   enabled: process.env.NODE_ENV !== "development" && !process.env.CI,
+
+  // Add exception metadata to the event
+  beforeSend(event, hint) {
+    const error = hint.originalException;
+    let nicified;
+    try {
+      nicified = nicify(error);
+    } catch (e) {
+      nicified = `Error occurred during nicification: ${e}`;
+    }
+    if (error instanceof Error) {
+      event.extra = {
+        ...event.extra,
+        cause: error.cause,
+        errorProps: {
+          ...error,
+        },
+        nicifiedError: nicify(error),
+      };
+    }
+    return event;
+  },
 });