+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @stackframe/stack-backend
 
+## 2.5.16
+
+### Patch Changes
+
+- @stackframe/stack-emails@2.5.16
+- @stackframe/stack-shared@2.5.16
+
+## 2.5.15
+
+### Patch Changes
+
+- Webhooks
+- Updated dependencies
+  - @stackframe/stack-emails@2.5.15
+  - @stackframe/stack-shared@2.5.15
+
 ## 2.5.14
 
 ### Patch Changes