+++ apps/e2e/CHANGELOG.md
@@ -1,5 +1,20 @@
 # e2e-tests
 
+## 2.5.16
+
+### Patch Changes
+
+- Update README
+  - @stackframe/stack-shared@2.5.16
+
+## 2.5.15
+
+### Patch Changes
+
+- Webhooks
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.15
+
 ## 2.5.14
 
 ### Patch Changes