+++ packages/stack/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack
 
+## 2.5.16
+
+### Patch Changes
+
+- @stackframe/stack-sc@2.5.16
+- @stackframe/stack-shared@2.5.16
+- @stackframe/stack-ui@2.5.16
+
+## 2.5.15
+
+### Patch Changes
+
+- Webhooks
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.15
+  - @stackframe/stack-sc@2.5.15
+  - @stackframe/stack-ui@2.5.15
+
 ## 2.5.14
 
 ### Patch Changes